world impl wip

+++ Core/Public/Engine/Engine/ECS/Types/Entity.h
@@ -1,3 +1,3 @@
 #pragma once
 
-using Entity = size_t;
+using Entity = uint32_t;
